refactor: amount int를 long으로 변경

+++ apps/dragons_api/src/main/java/com/dragons/application/payment/dto/PaymentTossCommand.java
@@ -1,7 +1,7 @@
 package com.dragons.application.payment.dto;
 
 public record PaymentTossCommand(
-    int amount,
+    long amount,
     String orderName,
     String customerName,
     String planType
